Move repo perms syncer to worker
Since we have distributed rate limits now, the last dependency is broken and we can move this subsystem around freely.
To make repo-updater more lightweight, worker will be the new home of this system.

## Test plan

Ran stack locally, CI still passes including integration tests.
