+++ cmd/repo-updater/shared/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
     deps = [
         "//cmd/frontend/envvar",
         "//cmd/frontend/globals",
-        "//cmd/repo-updater/internal/authz",
         "//cmd/repo-updater/internal/repoupdater",
         "//internal/actor",
         "//internal/api",
@@ -27,6 +26,7 @@ go_library(
         "//internal/conf/conftypes",
         "//internal/database",
         "//internal/database/connections/live",
+        "//internal/database/dbutil",
         "//internal/debugserver",
         "//internal/encryption/keyring",
         "//internal/env",
@@ -45,12 +45,12 @@ go_library(
         "//internal/repos/scheduler",
         "//internal/repoupdater/v1:repoupdater",
         "//internal/service",
-        "//internal/timeutil",
         "//internal/trace",
         "//internal/types",
         "//lib/errors",
         "@com_github_graph_gophers_graphql_go//relay",
         "@com_github_prometheus_client_golang//prometheus",
+        "@com_github_prometheus_client_golang//prometheus/promauto",
         "@com_github_sourcegraph_log//:log",
         "@io_opentelemetry_go_otel//:otel",
         "@org_golang_google_grpc//:go_default_library",
