+++ cmd/worker/internal/authz/BUILD.bazel
@@ -6,24 +6,31 @@ go_library(
     srcs = [
         "metrics.go",
         "perms_syncer.go",
+        "perms_syncer_job.go",
         "perms_syncer_worker.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/repo-updater/internal/authz",
-    visibility = ["//cmd/repo-updater:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/worker/internal/authz",
+    visibility = ["//cmd/worker:__subpackages__"],
     deps = [
         "//cmd/frontend/envvar",
+        "//cmd/worker/job",
+        "//cmd/worker/shared/init/db",
+        "//internal/actor",
         "//internal/api",
         "//internal/authz",
         "//internal/collections",
         "//internal/conf",
         "//internal/database",
         "//internal/database/basestore",
+        "//internal/env",
         "//internal/errcode",
         "//internal/extsvc",
         "//internal/extsvc/github",
         "//internal/featureflag",
+        "//internal/goroutine",
         "//internal/observation",
         "//internal/repos",
+        "//internal/timeutil",
         "//internal/trace",
         "//internal/types",
         "//internal/workerutil",
