+++ cmd/worker/internal/authz/testdata/vcr/TestIntegration_GitHubInternalRepositories.yaml
