+++ cmd/worker/internal/authz/testdata/vcr/TestIntegration_GitHubPermissions/user-centric/no-groups.yaml
