+++ cmd/worker/internal/authz/testdata/vcr/TestIntegration_GitLabPermissions/test_gitLabProjectVisibilityExperimental_feature_flag.yaml
