+++ cmd/worker/shared/main.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/globals"
 	"github.com/sourcegraph/sourcegraph/cmd/worker/internal/auth"
+	workerauthz "github.com/sourcegraph/sourcegraph/cmd/worker/internal/authz"
 	"github.com/sourcegraph/sourcegraph/cmd/worker/internal/batches"
 	"github.com/sourcegraph/sourcegraph/cmd/worker/internal/codeintel"
 	"github.com/sourcegraph/sourcegraph/cmd/worker/internal/codemonitors"
@@ -127,6 +128,8 @@ func LoadConfig(registerEnterpriseMigrators oobmigration.RegisterMigratorsFunc) 
 		"github-apps-installation-validation-job": githubapps.NewGitHubApsInstallationJob(),
 
 		"exhaustive-search-job": search.NewSearchJob(),
+
+		"repo-perms-syncer": workerauthz.NewPermsSyncerJob(),
 	}
 
 	var config Config
