+++ monitoring/definitions/worker.go
@@ -161,6 +161,8 @@ func Worker() *monitoring.Dashboard {
 			shared.CodeIntelligence.NewGitserverClientGroup(containerName),
 			shared.CodeIntelligence.NewDependencyReposStoreGroup(containerName),
 
+			repoPermsSyncerGroup(monitoring.ObservableOwnerSource),
+
 			shared.GitServer.NewClientGroup(containerName),
 
 			shared.Batches.NewDBStoreGroup(containerName),
@@ -269,6 +271,185 @@ func Worker() *monitoring.Dashboard {
 				HumanServiceName:    "worker",
 				InstanceFilterRegex: `${instance:regex}`,
 			}, monitoring.ObservableOwnerInfraOrg),
+		},
+	}
+}
+
+func repoPermsSyncerGroup(owner monitoring.ObservableOwner) monitoring.Group {
+	return monitoring.Group{
+		Title:  "Permissions",
+		Hidden: true,
+		Rows: []monitoring.Row{
+			{
+				{
+					Name:           "user_success_syncs_total",
+					Description:    "total number of user permissions syncs",
+					Query:          `sum(src_repo_perms_syncer_success_syncs{type="user"})`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the total number of user permissions sync completed.",
+				},
+				{
+					Name:           "user_success_syncs",
+					Description:    "number of user permissions syncs [5m]",
+					Query:          `sum(increase(src_repo_perms_syncer_success_syncs{type="user"}[5m]))`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the number of users permissions syncs completed.",
+				},
+				{
+					Name:           "user_initial_syncs",
+					Description:    "number of first user permissions syncs [5m]",
+					Query:          `sum(increase(src_repo_perms_syncer_initial_syncs{type="user"}[5m]))`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the number of permissions syncs done for the first time for the user.",
+				},
+			},
+			{
+
+				{
+					Name:           "repo_success_syncs_total",
+					Description:    "total number of repo permissions syncs",
+					Query:          `sum(src_repo_perms_syncer_success_syncs{type="repo"})`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the total number of repo permissions sync completed.",
+				},
+				{
+					Name:           "repo_success_syncs",
+					Description:    "number of repo permissions syncs over 5m",
+					Query:          `sum(increase(src_repo_perms_syncer_success_syncs{type="repo"}[5m]))`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the number of repos permissions syncs completed.",
+				},
+				{
+					Name:           "repo_initial_syncs",
+					Description:    "number of first repo permissions syncs over 5m",
+					Query:          `sum(increase(src_repo_perms_syncer_initial_syncs{type="repo"}[5m]))`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the number of permissions syncs done for the first time for the repo.",
+				},
+			},
+			{
+				{
+					Name:           "users_consecutive_sync_delay",
+					Description:    "max duration between two consecutive permissions sync for user",
+					Query:          `max(max_over_time (src_repo_perms_syncer_perms_consecutive_sync_delay{type="user"} [1m]))`,
+					Panel:          monitoring.Panel().LegendFormat("seconds").Unit(monitoring.Seconds),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the max delay between two consecutive permissions sync for a user during the period.",
+				},
+				{
+					Name:           "repos_consecutive_sync_delay",
+					Description:    "max duration between two consecutive permissions sync for repo",
+					Query:          `max(max_over_time (src_repo_perms_syncer_perms_consecutive_sync_delay{type="repo"} [1m]))`,
+					Panel:          monitoring.Panel().LegendFormat("seconds").Unit(monitoring.Seconds),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the max delay between two consecutive permissions sync for a repo during the period.",
+				},
+			},
+			{
+				{
+					Name:           "users_first_sync_delay",
+					Description:    "max duration between user creation and first permissions sync",
+					Query:          `max(max_over_time(src_repo_perms_syncer_perms_first_sync_delay{type="user"}[1m]))`,
+					Panel:          monitoring.Panel().LegendFormat("seconds").Unit(monitoring.Seconds),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the max delay between user creation and their permissions sync",
+				},
+				{
+					Name:           "repos_first_sync_delay",
+					Description:    "max duration between repo creation and first permissions sync over 1m",
+					Query:          `max(max_over_time(src_repo_perms_syncer_perms_first_sync_delay{type="repo"}[1m]))`,
+					Panel:          monitoring.Panel().LegendFormat("seconds").Unit(monitoring.Seconds),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the max delay between repo creation and their permissions sync",
+				},
+			},
+			{
+				{
+					Name:           "permissions_found_count",
+					Description:    "number of permissions found during user/repo permissions sync",
+					Query:          `sum by (type) (src_repo_perms_syncer_perms_found)`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the number permissions found during users/repos permissions sync.",
+				},
+				{
+					Name:           "permissions_found_avg",
+					Description:    "average number of permissions found during permissions sync per user/repo",
+					Query:          `avg by (type) (src_repo_perms_syncer_perms_found)`,
+					Panel:          monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:          owner,
+					NoAlert:        true,
+					Interpretation: "Indicates the average number permissions found during permissions sync per user/repo.",
+				},
+			},
+			{
+				{
+					Name:        "perms_syncer_outdated_perms",
+					Description: "number of entities with outdated permissions",
+					Query:       `max by (type) (src_repo_perms_syncer_outdated_perms)`,
+					Warning:     monitoring.Alert().GreaterOrEqual(100).For(5 * time.Minute),
+					Panel:       monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:       owner,
+					NextSteps: `
+						- **Enabled permissions for the first time:** Wait for few minutes and see if the number goes down.
+						- **Otherwise:** Increase the API rate limit to [GitHub](https://docs.sourcegraph.com/admin/external_service/github#github-com-rate-limits), [GitLab](https://docs.sourcegraph.com/admin/external_service/gitlab#internal-rate-limits) or [Bitbucket Server](https://docs.sourcegraph.com/admin/external_service/bitbucket_server#internal-rate-limits).
+					`,
+				},
+			},
+			{
+				{
+					Name:        "perms_syncer_sync_duration",
+					Description: "95th permissions sync duration",
+					Query:       `histogram_quantile(0.95, max by (le, type) (rate(src_repo_perms_syncer_sync_duration_seconds_bucket[1m])))`,
+					Warning:     monitoring.Alert().GreaterOrEqual(30).For(5 * time.Minute),
+					Panel:       monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Seconds),
+					Owner:       owner,
+					NextSteps:   "Check the network latency is reasonable (<50ms) between the Sourcegraph and the code host.",
+				},
+			},
+			{
+				{
+					Name:        "perms_syncer_sync_errors",
+					Description: "permissions sync error rate",
+					Query:       `max by (type) (ceil(rate(src_repo_perms_syncer_sync_errors_total[1m])))`,
+					Critical:    monitoring.Alert().GreaterOrEqual(1).For(time.Minute),
+					Panel:       monitoring.Panel().LegendFormat("{{type}}").Unit(monitoring.Number),
+					Owner:       owner,
+					NextSteps: `
+						- Check the network connectivity the Sourcegraph and the code host.
+						- Check if API rate limit quota is exhausted on the code host.
+					`,
+				},
+				{
+					Name:        "perms_syncer_scheduled_repos_total",
+					Description: "total number of repos scheduled for permissions sync",
+					Query:       `max(rate(src_repo_perms_syncer_schedule_repos_total[1m]))`,
+					NoAlert:     true,
+					Panel:       monitoring.Panel().Unit(monitoring.Number),
+					Owner:       owner,
+					Interpretation: `
+						Indicates how many repositories have been scheduled for a permissions sync.
+						More about repository permissions synchronization [here](https://docs.sourcegraph.com/admin/permissions/syncing#scheduling)
+					`,
+				},
+			},
 		},
 	}
 }
